Replaced request with node fetch
Decided to do the replace because node fetch supports promises natively where as request did not use promises at all without the need of another library.

Had to change the struture of the queries to match the fetch api.

Also decided to use use async/await to learn more and see if it cleans the code up. Definitely helps make it easier to read. They are essentialy nice syntax sugar around promises because they ulimately return a promise.